pyproject.toml
[tool.poetry]
name = "techstormhack"
version = "0.1.0"
description = ""
authors = ["werserk <werserk@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
llvmlite = "^0.43.0"
numba = "^0.60.0"
pyannote-audio = "^3.3.1"
numpy = "1.26.0"
sounddevice = "^0.5.0"
realtimestt = "^0.2.41"
dlib = "^19.24.6"
face-recognition = "^1.3.0"
opencv-python = "^4.10.0.84"
ffmpeg-python = "^0.2.0"
resampy = "^0.4.3"
vosk = "^0.3.45"
transformers = "^4.44.2"
pydub = "^0.25.1"
moviepy = "^1.0.3"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
